requirements.txt
paramiko==3.4.0
boto3==1.34.69
python-magic==0.4.27 
python-dotenv==1.1.0
cryptography==41.0.7
cffi==1.17.0
pycparser==2.21
six==1.16.0